Add a basic JS challenge to the captcha wall
The server will generate a "correct" string and hash it.
It will then generate a number of other strings.
It will send all the strings to the client.
It will also send a hash of the correct string.

The client then has to hash the strings until it finds the correct string.
The correct string will then be sent along with the form request.
The string will be checked against the stored correct string on server.
If it matches it'll pass.